Moved folders into src
