+++ vynl/src/app/api/auth/confirm+api.ts
@@ -3,6 +3,7 @@ import type { SerializeOptions } from "cookie";
 import { type EmailOtpType } from '@supabase/supabase-js';
 
 export async function GET(req: Request) {
+    console.log('HERE SERVER');
     const url = new URL(req.url);
     const token_hash = url.searchParams.get('token_hash');
     const type = url.searchParams.get('type') as EmailOtpType | null;
@@ -26,14 +27,14 @@ export async function GET(req: Request) {
                 },
                 },
             }
-    )
-    const { error } = await supabase.auth.verifyOtp({
-        type,
-        token_hash,
-    })
-    if (!error) {
-        return Response.redirect(`/${next.slice(1)}`, 303);
-    }
+        )
+        const { error } = await supabase.auth.verifyOtp({
+            type,
+            token_hash,
+        })
+        if (!error) {
+            return Response.redirect(`/${next.slice(1)}`, 303);
+        }
     }
     // return the user to an error page with some instructions
     return Response.redirect('/auth/auth-code-error', 303);
