+++ vynl/src/components/external-link.tsx
