+++ vynl/src/components/haptic-tab.tsx
