+++ vynl/src/components/ui/icon-symbol.ios.tsx
