+++ vynl/src/components/ui/icon-symbol.tsx
